+++ server/src/main/java/com/emosaac/server/repository/genre/GenreRepository.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 @Repository
 public interface GenreRepository extends JpaRepository<Genre, Long> {
     @Query("select g from Genre g where  g.gerneId = :genreId")
-    Optional<Genre> findById(@Param("genreId") Long genreId);
+    Optional<Genre> findById(@Param("genreId") Integer genreId);
 
     @Query("select g from Genre g where substring(g.gerneId, 1, 1) = '1'")
     List<Genre> findWebtoon();
